+++ compiler-tests/src/test/data/diagnostic/provides/BindsWithBodiesShouldBePrivate_InInterface.kt
@@ -1,5 +1,6 @@
 // RENDER_DIAGNOSTICS_FULL_TEXT
 // PUBLIC_PROVIDER_SEVERITY: WARN
+// DISABLE_TRANSFORM_PROVIDERS_TO_PRIVATE
 
 interface ExampleGraph {
   @Binds val String.provideCharSequence: CharSequence get() = this