+++ compiler-tests/src/test/data/diagnostic/provides/Binds_interface_ShouldNotHaveBodies.kt
@@ -1,5 +1,6 @@
 // RENDER_DIAGNOSTICS_FULL_TEXT
 // PUBLIC_PROVIDER_SEVERITY: WARN
+// DISABLE_TRANSFORM_PROVIDERS_TO_PRIVATE
 
 interface ExampleGraph {
   @Binds val Int.bind: Number get() = this