+++ compiler-tests/src/test/data/diagnostic/provides/PrivateProviderOption_Error.kt
@@ -1,5 +1,6 @@
 // RENDER_DIAGNOSTICS_FULL_TEXT
 // PUBLIC_PROVIDER_SEVERITY: ERROR
+// DISABLE_TRANSFORM_PROVIDERS_TO_PRIVATE
 
 interface ExampleGraph {
   @Provides val provideCharSequence: String get() = "Hello"