+++ compiler-tests/src/test/data/diagnostic/provides/PrivateProviderOption_None.kt
@@ -1,4 +1,5 @@
 // PUBLIC_PROVIDER_SEVERITY: NONE
+// DISABLE_TRANSFORM_PROVIDERS_TO_PRIVATE
 
 interface ExampleGraph {
   @Provides val provideCharSequence: String get() = "Hello"