+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/MetroCompilerTest.kt
@@ -101,6 +101,8 @@ abstract class MetroCompilerTest {
                 processor.option(entry.raw.cliOption, enableTopLevelFunctionInjection)
               MetroOption.GENERATE_HINT_PROPERTIES ->
                 processor.option(entry.raw.cliOption, generateHintProperties)
+              MetroOption.TRANSFORM_PROVIDERS_TO_PRIVATE ->
+                processor.option(entry.raw.cliOption, transformProvidersToPrivate)
               MetroOption.PUBLIC_PROVIDER_SEVERITY ->
                 processor.option(entry.raw.cliOption, publicProviderSeverity)
               MetroOption.LOGGING -> {