+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/transformers/DependencyGraphTransformerTest.kt
@@ -449,6 +449,7 @@ class DependencyGraphTransformerTest : MetroCompilerTest() {
             .trimIndent()
         ),
         expectedExitCode = ExitCode.COMPILATION_ERROR,
+        options = metroOptions.copy(transformProvidersToPrivate = false),
       )
 
     result.assertDiagnostics(